Another WIP checkin point
Using gltfjsx to generate React FC for us instead of doing it by
hand, looks much easier to reason about with and takes care
of the loading issues originally exhibited.

Remaining work is to use react-three-flex to fix the positioning
problem that I thought would be a huge issue and tons of math.
Other people did it already, we can come back to it later.